Add files via upload
Added a line to change out all the 0s at the end of time/temperature vectors in the Temps array to NaN (it might accidentally remove the last datapoint, so we need to go over it, but it does allow the plots to work).

I don't think the simulation is running properly. I ran it for a month and it was still periodic instead of reaching equilibrium. It also looks like the insulation has no effect, so that would explain the problem.